+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Unreleased
+
+- Add `@client.Record.build(attributes)`
+- Add `Record#persisted?`
 - Implement a custom User-Agent
 
 ## 0.0.1 (2014-03-07)